Properly handle failure to bind to address+port
Throwing in an event listener will **not** cause the outer function to throw an error since Listener.listen is not the caller. The error is actually passed on to whoever called `emitter.emit("error", error)`, but since socket.emit is an async method (callback style), it turns into an unhandled promise rejection.
To prevent that, the error event and the `socket.bind` call has been wrapped in a new Promise and to properly pass the error off to the caller, should it fail (e.g. port in use), and will resolve once successfully connected along with removing the error event listener as it's no longer needed afterward.

Simpler Reproducible code:
```js
const listener = new (events.EventEmitter)()
//In the actual code, socket.bind is an async method, 
async function setup() {
    await new Promise(res => setTimeout(res, 1000))
    listener.emit("error", new Error("Uh-oh"))
}
function init() {
    listener.on("error", e => {
        throw e
    })
    setup()
}